+++ MarkdownMonster/_Classes/MarkdownParser/MarkdownParserMarkDig.cs
@@ -132,10 +132,10 @@ namespace MarkdownMonster
                 builder = builder.UseTaskLists();
             if (options.SmartyPants)
                 builder = builder.UseSmartyPants();
-            if(options.Diagrams)
-                builder = builder.UseDiagrams();
-            if (options.Mathmatics)
-                builder = builder.UseMathematics();
+            //if(options.Diagrams)
+            //    builder = builder.UseDiagrams();
+            //if (options.Mathmatics)
+            //    builder = builder.UseMathematics();
             if (options.CustomContainers)
                 builder = builder.UseCustomContainers();
             if (options.Attributes)